fix: resolve CI failures for Hytte-zptt
Extract computeWeakestFacts and findUserRank into mathSummaryUtils.ts
to fix react-refresh/only-export-components lint errors — files that
export React components must not also export non-component functions.
